requirements.txt
asgiref==3.2.10
boto3==1.14.56
botocore==1.17.56
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-appconf==1.0.4
django-ckeditor==5.9.0
django-embed-video==1.3.3
django-heroku==0.3.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-markdownx==3.0.1
django-mdeditor==0.1.18
django-storages==1.10
docutils==0.15.2
ExifRead==2.3.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
jmespath==0.10.0
Markdown==3.2.2
oauthlib==3.1.0
pilkit==2.0
Pillow==7.2.0
psycopg2==2.8.5
Pygments==2.6.1
pymdown-extensions==7.1
PyMySQL==0.10.0
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sentry-sdk==0.17.4
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
zipp==3.1.0
